extract logic to util function
